+++ test_utils.py
@@ -497,6 +497,7 @@ def test_create_app_function():
         app_path = create_app(unzip='*', pip_args=['six'], lazy_install=True)
         subprocess.Popen([sys.executable,
                           str(app_path), '--activate-zipapps']).wait()
+        assert Path('app.pyz').stat().st_mode != 33279
         for _path in Path('zipapps_cache/app').rglob('*'):
             if _path.name == 'six.py':
                 assert _path.stat().st_mode != 33279, _path.stat().st_mode
@@ -509,6 +510,7 @@ def test_create_app_function():
                               chmod='777')
         subprocess.Popen([sys.executable,
                           str(app_path), '--activate-zipapps']).wait()
+        assert Path('app.pyz').stat().st_mode == 33279
         for _path in Path('zipapps_cache/app').rglob('*'):
             if _path.name == 'six.py':
                 assert _path.stat().st_mode == 33279, _path.stat().st_mode